requirements.txt
asgiref
certifi
cffi
charset-normalizer
cryptography
defusedxml
dj-database-url
Django
django-cors-headers
django-taggit
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
idna
oauthlib
pillow
psycopg2
psycopg2-binary
pycparser
PyJWT
python-decouple
python3-openid
requests
requests-oauthlib
social-auth-app-django
social-auth-core
sqlparse
typing_extensions
tzdata
urllib3
whitenoise
